Some basic auth working now
Only implemented the authenticate method, and it is using hard coded
parameters instead of params from the form.

Ember CLI is able to proxy to the rails server using this startup
command: ember serve --proxy http://localhost:3000

Implementing restore is the next most important thing to do.

Some indicators dded to application.hbs to show whether or not we are
logged in.