Hook up proper navbar interactions
Import bootstrap collapse so navbar actually works when collapsed.

Hook up custom event handling in the application view to hide the navbar
when any link in the navbar is clicked.  Bootstrap does not do this,
probably because they expect links to do full reloads.  It will would
cleaner to have a subview for the navbar instead of handling this at the
application level.