Super hack to get bootstrap icons working
This is a hack because we use most of bootstrap via bower, but we check
in the icons.  Ideally these icons would be copied into place as part of
the build.

It looks like we could do this via regular import if we updated our
ember-cli as they added a feature for this in 0.0.28