Sort notes by updated_at on search index
It may make more sense to use an ArrayController instead of an
ObjectController, but it was worth figuring out how to do this without
making the switch.

Now that we learned how to do this on a arbitrary property of an
ObjectController we can switch to an ArrayController if it is still
applicable when we have time to do it.